--- lib/sidekiq/cleaner/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Sidekiq
-  module Cleaner
-    VERSION = "0.3.6"
-  end
-end
