--- lib/sidekiq/cleaner/views/cleaner_failures.erb
@@ -1,70 +1,0 @@
-<header class="row head">
-  <div class="col-sm-12">
-    <h3><%= "<b>#{@total_size}</b> failures in <b>#{@req_job_class}</b> queue with <b>#{@req_error_class}</b> exception in <b>#{@req_bucket_name}</b>"  %></h3>
-  </div>
-  <div class="col-sm-12">
-      <% if @dead_jobs.size > 0 && @total_size > @count %>
-          <div class="col-sm-4">
-            <%= erb :_paging, :locals => { :url => "#{root_path}" + @cleaner_path } %>
-          </div>
-      <% end %>
-  </div>
-  <%= filtering('dead') %>
-</header>
-
-<% if @dead_jobs.size > 0 %>
-    <form action="<%= root_path %>cleaner" method="post">
-      <%= respond_to?(:csrf_tag) && csrf_tag %>
-      <table class="table table-striped table-bordered table-white">
-        <thead>
-          <tr>
-            <th width="20px" class="table-checkbox">
-              <label>
-                <input type="checkbox" class="check_all" />
-              </label>
-            </th>
-            <th width="25%"><%= t('LastRetry') %></th>
-            <th><%= t('Queue') %></th>
-            <th><%= t('Worker') %></th>
-            <th><%= t('Arguments') %></th>
-            <th><%= t('Error') %></th>
-          </tr>
-        </thead>
-        <% @dead_jobs.each do |entry| %>
-            <tr>
-              <td class="table-checkbox">
-                <label>
-                  <input type='checkbox' name='key[]' value='<%= job_params(entry.item, entry.score) %>' />
-                </label>
-              </td>
-              <td>
-                <a href="<%= root_path %>morgue/<%= job_params(entry.item, entry.score) %>"><%= relative_time(entry.at) %></a>
-              </td>
-              <td>
-                <a href="<%= root_path %>queues/<%= entry.queue %>"><%= entry.queue %></a>
-              </td>
-              <td><%= entry.klass %></td>
-              <td>
-                <div class="args"><%= display_args(entry.args) %></div>
-              </td>
-              <td>
-                <div><%= h truncate("#{entry['error_class']}: #{entry['error_message']}", 200) %></div>
-              </td>
-            </tr>
-        <% end %>
-      </table>
-      <input class="btn btn-primary btn-xs pull-left" type="submit" name="retry" value="<%= t('RetryNow') %>" />
-      <input class="btn btn-danger btn-xs pull-left" type="submit" name="delete" value="<%= t('Delete') %>" />
-    </form>
-
-    <form action="<%= root_path %>cleaner/<%= @req_job_class %>/<%= @req_error_class %>/<%= @req_bucket_name %>/delete" method="post">
-      <%= respond_to?(:csrf_tag) && csrf_tag %>
-      <input class="btn btn-danger btn-xs pull-right" type="submit" name="delete" value="<%= t('DeleteAll') %>" data-confirm="<%= t('AreYouSure') %>" />
-    </form>
-    <form action="<%= root_path %>cleaner/<%= @req_job_class %>/<%= @req_error_class %>/<%= @req_bucket_name %>/retry" method="post">
-      <%= respond_to?(:csrf_tag) && csrf_tag %>
-      <input class="btn btn-danger btn-xs pull-right" type="submit" name="retry" value="<%= t('RetryAll') %>" data-confirm="<%= t('AreYouSure') %>" />
-    </form>
-<% else %>
-    <div class="alert alert-success"><%= t('NoDeadJobsFound') %></div>
-<% end %>
