+++ lib/sidekiq/undertaker/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Sidekiq
+  module Undertaker
+    VERSION = "1.0.0.rc01"
+  end
+end
