--- spec/fixtures/approvals/sidekiq_cleaner_webextension/show/when_overview_page_is_called/behaves_like_a_cleaner_page/the_displayed_page_is_correct.approved.txt
@@ -1,234 +1,0 @@
-<!doctype html>
-<html dir="ltr">
-  <head>
-    <title>[TEST] Sidekiq</title>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width,initial-scale=1.0" />
-
-    <link href="/stylesheets/bootstrap.css" media="screen" rel="stylesheet" type="text/css" />
-    
-
-    <link href="/stylesheets/application.css" media="screen" rel="stylesheet" type="text/css" />
-    
-
-    <link rel="shortcut icon" type="image/ico" href="/images/favicon.ico" />
-    <script type="text/javascript" src="/javascripts/application.js"></script>
-    <meta name="google" content="notranslate" />
-    
-  </head>
-  <body class="admin" data-poll-path="" data-locale="en">
-    <div class="navbar navbar-default navbar-fixed-top">
-  <div class="container-fluid">
-    <div class="navbar-header" data-navbar="static">
-      <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar-menu">
-        <span class="icon-bar"></span>
-        <span class="icon-bar"></span>
-        <span class="icon-bar"></span>
-      </button>
-      <div class="navbar-toggle collapsed navbar-livereload">
-        
-  
-    <a id="live-poll" class="btn btn-primary" href="/cleaner?poll=true">Live Poll</a>
-  
-
-
-        
-      </div>
-      <a class="navbar-brand" href="/">
-        Sidekiq
-        <span class="status">
-  <i class="status-sprite status-idle"></i>
-  idle
-</span>
-
-      </a>
-    </div>
-
-    <div class="collapse navbar-collapse" id="navbar-menu">
-      <ul class="nav navbar-nav" data-navbar="static">
-        
-          
-            <li class="">
-            <a href="/">Dashboard</a>
-            </li>
-          
-        
-          
-            <li class="">
-            <a href="/busy">Busy</a>
-            </li>
-          
-        
-          
-            <li class="">
-            <a href="/queues">Queues</a>
-            </li>
-          
-        
-          
-            <li class="">
-            <a href="/retries">Retries</a>
-            </li>
-          
-        
-          
-            <li class="">
-            <a href="/scheduled">Scheduled</a>
-            </li>
-          
-        
-          
-            <li class="">
-            <a href="/morgue">Dead</a>
-            </li>
-          
-        
-
-        
-          <li class="active" data-navbar="custom-tab">
-            <a href="/cleaner">Cleaner</a>
-          </li>
-        
-
-        <li class="navbar-livereload">
-          <div class="poll-wrapper">
-            
-  
-    <a id="live-poll" class="btn btn-primary" href="/cleaner?poll=true">Live Poll</a>
-  
-
-
-            
-          </div>
-        </li>
-      </ul>
-    </div>
-  </div>
-</div>
-
-    <div id="page">
-      <div class="container">
-        <div class="row">
-          <div class="col-sm-12 summary_bar">
-            <ul class="list-unstyled summary row">
-  <li class="processed col-sm-1">
-    <span class="count">0</span>
-    <span class="desc">Processed</span>
-  </li>
-  <li class="failed col-sm-1">
-    <span class="count">0</span>
-    <span class="desc">Failed</span>
-  </li>
-  <li class="busy col-sm-1">
-    <a href="/busy">
-      <span class="count">0</span>
-      <span class="desc">Busy</span>
-    </a>
-  </li>
-  <li class="enqueued col-sm-1">
-    <a href="/queues">
-      <span class="count">0</span>
-      <span class="desc">Enqueued</span>
-    </a>
-  </li>
-  <li class="retries col-sm-1">
-    <a href="/retries">
-      <span class="count">0</span>
-      <span class="desc">Retries</span>
-    </a>
-  </li>
-  <li class="scheduled col-sm-1">
-    <a href="/scheduled">
-      <span class="count">0</span>
-      <span class="desc">Scheduled</span>
-    </a>
-  </li>
-  <li class="dead col-sm-1">
-    <a href="/morgue">
-      <span class="count">4</span>
-      <span class="desc">Dead</span>
-    </a>
-  </li>
-</ul>
-
-          </div>
-
-          <div class="col-sm-12">
-            <header class="row header">
-  <div class="col-sm-12">
-      <h3><b>4</b> failures </h3>
-  </div>
-</header>
-
-<table class="table table-striped table-bordered table-white">
-  <thead>
-    <tr>
-      <th style="width: 20%">Failure</th>
-      <th style="width: 10%">Failed</th>
-      <th style="width: 10%">1 hour</th>
-      <th style="width: 10%">3 hours</th>
-      <th style="width: 10%">1 day</th>
-      <th style="width: 10%">3 days</th>
-      <th style="width: 10%">7 days</th>
-    </tr>
-  </thead>
-  
-      <tr>
-        <td><a href='/cleaner/AllErrors/AllErrors/total_failures'>AllErrors</a></td>
-        <td><a href='/cleaner/AllErrors/total_failures'>4</a></td>
-        <td><a href='/cleaner/AllErrors/1_hour'>4</a></td>
-        <td><a href='/cleaner/AllErrors/3_hours'>0</a></td>
-        <td><a href='/cleaner/AllErrors/1_day'>0</a></td>
-        <td><a href='/cleaner/AllErrors/3_days'>0</a></td>
-        <td><a href='/cleaner/AllErrors/7_days'>0</a></td>
-      </tr>
-  
-      <tr>
-        <td><a href='/cleaner/HardWorker/AllErrors/total_failures'>HardWorker</a></td>
-        <td><a href='/cleaner/HardWorker/total_failures'>3</a></td>
-        <td><a href='/cleaner/HardWorker/1_hour'>3</a></td>
-        <td><a href='/cleaner/HardWorker/3_hours'>0</a></td>
-        <td><a href='/cleaner/HardWorker/1_day'>0</a></td>
-        <td><a href='/cleaner/HardWorker/3_days'>0</a></td>
-        <td><a href='/cleaner/HardWorker/7_days'>0</a></td>
-      </tr>
-  
-      <tr>
-        <td><a href='/cleaner/HardWorker1/AllErrors/total_failures'>HardWorker1</a></td>
-        <td><a href='/cleaner/HardWorker1/total_failures'>1</a></td>
-        <td><a href='/cleaner/HardWorker1/1_hour'>1</a></td>
-        <td><a href='/cleaner/HardWorker1/3_hours'>0</a></td>
-        <td><a href='/cleaner/HardWorker1/1_day'>0</a></td>
-        <td><a href='/cleaner/HardWorker1/3_days'>0</a></td>
-        <td><a href='/cleaner/HardWorker1/7_days'>0</a></td>
-      </tr>
-  
-</table>
-
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="navbar navbar-fixed-bottom navbar-inverse ltr">
-  <div class="navbar-inner">
-    <div class="container text-center">
-        <ul class="nav">
-          <li>
-            <p class="navbar-text product-version">Sidekiq v*EXCLUDED*</p>
-          </li>
-          <li>
-            <p class="navbar-text redis-url" title="redis://127.0.0.1:6379/0">redis://127.0.0.1:6379/0</p>
-          </li>
-          <li>
-            <p class="navbar-text server-utc-time">20:57:00 UTC</p>
-          </li>
-          <li>
-            <p class="navbar-text"><a style="color: gray;" href="https://github.com/mperham/sidekiq/wiki">docs</a></p>
-          </li>
-        </ul>
-    </div>
-  </div>
-</div>
-
-  </body>
-</html>
