+++ spec/spec_helper.rb
@@ -30,7 +30,7 @@ if ENV["COVERAGE"]
   end
 end
 
-require "sidekiq/cleaner"
+require "sidekiq/undertaker"
 
 RSpec.configure do |config|
   config.run_all_when_everything_filtered = true
